third_party/blink/web_tests/fast/animation/request-animation-frame-iframe2-expected.txt

Tests requestAnimationFrame in both an iframe and main frame

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS callbackInvoked is true
PASS mainFrameCallbackInvoked is true
PASS successfullyParsed is true

TEST COMPLETE
